# static/swagger.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Task Manager API",
        "description": "API REST pour la gestion de tâches. Projet DevOps Académique.",
        "version": "1.0.0",
        "contact": {
            "name": "Nour Moussi",
            "url": "https://github.com/NourMoussi"
        }
    },
    "servers": [
        {
            "url": "http://localhost:5000",
            "description": "Serveur de développement local"
        }
    ],
    "tags": [
        {
            "name": "Tasks",
            "description": "Opérations sur les tâches"
        },
        {
            "name": "System",
            "description": "Opérations système (Health check, Metrics)"
        }
    ],
    "paths": {
        "/health": {
            "get": {
                "tags": [
                    "System"
                ],
                "summary": "Vérification de l'état du service",
                "description": "Retourne l'état de santé de l'application.",
                "responses": {
                    "200": {
                        "description": "Service en ligne",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "properties": {
                                        "status": {
                                            "type": "string",
                                            "example": "healthy"
                                        },
                                        "version": {
                                            "type": "string",
                                            "example": "1.0.0"
                                        },
                                        "timestamp": {
                                            "type": "string",
                                            "format": "date-time"
                                        }
                                    }
                                }
                            }
                        }
                    }
                }
            }
        },
        "/metrics": {
            "get": {
                "tags": [
                    "System"
                ],
                "summary": "Métriques Prometheus",
                "description": "Expose les métriques applicatives au format Prometheus.",
                "responses": {
                    "200": {
                        "description": "Métriques textuelles",
                        "content": {
                            "text/plain": {}
                        }
                    }
                }
            }
        },
        "/tasks": {
            "get": {
                "tags": [
                    "Tasks"
                ],
                "summary": "Lister toutes les tâches",
                "responses": {
                    "200": {
                        "description": "Liste des tâches récupérée avec succès",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/Task"
                                    }
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "Tasks"
                ],
                "summary": "Créer une nouvelle tâche",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/TaskInput"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Tâche créée",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Task"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Données invalides"
                    },
                    "422": {
                        "description": "Erreur de validation (champs manquants ou incorrects)"
                    }
                }
            }
        },
        "/tasks/{taskId}": {
            "get": {
                "tags": [
                    "Tasks"
                ],
                "summary": "Récupérer une tâche par ID",
                "parameters": [
                    {
                        "name": "taskId",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Tâche trouvée",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Task"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Tâche non trouvée"
                    }
                }
            },
            "put": {
                "tags": [
                    "Tasks"
                ],
                "summary": "Mettre à jour une tâche",
                "parameters": [
                    {
                        "name": "taskId",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/TaskInput"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Tâche mise à jour",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Task"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Tâche non trouvée"
                    }
                }
            },
            "delete": {
                "tags": [
                    "Tasks"
                ],
                "summary": "Supprimer une tâche",
                "parameters": [
                    {
                        "name": "taskId",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Tâche supprimée"
                    },
                    "404": {
                        "description": "Tâche non trouvée"
                    }
                }
            }
        }
    },
    "components": {
        "schemas": {
            "Task": {
                "type": "object",
                "properties": {
                    "id": {
                        "type": "string",
                        "format": "uuid",
                        "example": "a1b2c3d4-e5f6-..."
                    },
                    "title": {
                        "type": "string",
                        "example": "Faire les courses"
                    },
                    "description": {
                        "type": "string",
                        "example": "Acheter du lait et du pain"
                    },
                    "status": {
                        "type": "string",
                        "enum": [
                            "TODO",
                            "IN_PROGRESS",
                            "DONE"
                        ],
                        "example": "TODO"
                    },
                    "created_at": {
                        "type": "string",
                        "format": "date-time"
                    }
                }
            },
            "TaskInput": {
                "type": "object",
                "required": [
                    "title"
                ],
                "properties": {
                    "title": {
                        "type": "string"
                    },
                    "description": {
                        "type": "string"
                    },
                    "status": {
                        "type": "string",
                        "enum": [
                            "TODO",
                            "IN_PROGRESS",
                            "DONE"
                        ]
                    }
                }
            }
        }
    }
}